[wip][jit] big cpp test reorg
