+++ caffe2/CMakeLists.txt
@@ -466,7 +466,6 @@ if (NOT INTERN_BUILD_MOBILE OR NOT BUILD_CAFFE2_MOBILE)
       ${TORCH_SRC_DIR}/csrc/distributed/rpc/ScriptCall.cpp
       ${TORCH_SRC_DIR}/csrc/distributed/rpc/ScriptRet.cpp
       ${TORCH_SRC_DIR}/csrc/jit/export.cpp
-      ${TORCH_ROOT}/test/cpp/jit/test.cpp
     )
     if (NOT WIN32)
       list(APPEND TORCH_SRCS
