+++ test/cpp/jit/CMakeLists.txt
@@ -1,12 +1,15 @@
 set(JIT_TEST_ROOT ${TORCH_ROOT}/test/cpp/jit)
 
+file(GLOB JIT_TEST_SRCS ${JIT_TEST_ROOT}/test_*.cpp)
+set(JIT_TEST_SRCS ${JIT_TEST_SRCS} PARENT_SCOPE)
+
 add_executable(test_jit
   ${TORCH_ROOT}/test/cpp/common/main.cpp
-  ${JIT_TEST_ROOT}/test.cpp)
+  ${JIT_TEST_ROOT}/gtest.cpp
+  ${JIT_TEST_SRCS})
 
 target_link_libraries(test_jit PRIVATE torch gtest)
 target_include_directories(test_jit PRIVATE ${ATen_CPU_INCLUDE})
-target_compile_definitions(test_jit PRIVATE USE_GTEST)
 
 if (USE_CUDA)
   target_link_libraries(test_jit PRIVATE
