+++ tools/build_variables.py
@@ -439,6 +439,7 @@ def add_torch_libs():
         deps=[
             ":torch-cpp-cpu",
             ":thnn",
+            "//caffe2/test/cpp/jit:test",
             "//caffe2/torch/fb/init:init",
             "//caffe2/torch/lib/c10d:c10d_cpu",
             "//caffe2/torch/lib/libshm:libshm",
@@ -459,6 +460,7 @@ def add_torch_libs():
         deps=[
             ":torch-cpp-cuda",
             ":thnn",
+            "//caffe2/test/cpp/jit:test",
             "//caffe2/torch/fb/init:init",
             "//caffe2/torch/lib/c10d:c10d",
             "//caffe2/torch/lib/libshm:libshm",
