+++ torch/csrc/jit/testing/file_check.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <torch/csrc/WindowsTorchApiMacro.h>
+#include <string>
+#include <memory>
 
 namespace torch {
 namespace jit {
